Find usages of overridden methods
Summary:
```
class A { bar(): void {} }
class B extends A { bar(): void {} }
const a: A = new B();
a.bar();
```
As demonstrated above, a reference to a superclass method can end up calling a subclass method at runtime. Therefore, it is desirable to include calls to a superclass's method when looking for references to a subclass's method. The definition site of the superclass method is included for good measure.

It may be desirable to make this configurable. There are certain cases where this is probably not what the user is interested in. Right now, Nuclide does not have any UI which would allow us to differentiate between different types of find-refs requests but I think it would be worthwhile to build.

Using this architecture it would be easy to add flags to configure this behavior. It would also be easy to add a mode that includes subclasses as well.

Reviewed By: panagosg7

Differential Revision: D6697422

fbshipit-source-id: 1c7668b4a018c17f3d06db5512afc3d687484754