Add use_op to appropriate EIncompatible errors
Summary:
This diff adds a `use_op` to `EIncompatible` errors since many `use_t`s also have a `use_op`. In this diff I also added all other `use_t`s to a `util_use_op_of_use_t` function so that we remember to keep it updated.

For an example where this change is important for error positioning, see the changes I made to `type-destructors/errors.js`. You'll notice that two errors (the ones that use `$PropertyType`) don't have the correct position. This is because incompatible `GetPropT`s have their own error. `EIncompatibleGetProp` which is not `EIncompatible`. In a future diff I remove `EIncompatibleGetProp` to fix this.

Reviewed By: gabelevi

Differential Revision: D6628856

fbshipit-source-id: e51f76874b7b782d11c3815e1ee825eb59d39147