Add the friendly error type
Summary:
This is an infrastructure diff that adds the friendly error message format which I use in future diffs to improve Flow’s error messages. On its own, this diff should not change any of our error messages. This diff also does not add the CLI rendering for friendly errors. I do that in the next diff.

So what makes a friendly error message different from our current error message format? (Which I will from now on call the “classic error format.”) The classic error format is a collection of locations and sentence fragments. A friendly error message, on the other hand is made out of three main parts:

1. A primary location. It is _incredibly_ important that this location be the root cause of the error. Technically, Gabe defined this well: “a primary location in file A.js that mentions files X.js, Y.js, Z.js MUST have A.js as a direct or transitive dependent of X.js, Y.js and Z.js.” This constraint is important not just for usability, but also for our suppression comment strategy. This is the location we underline in the editor, and the only location we render inline for the CLI. Before we would render all of the locations which contribute to an error which adds noise.
2. A single error message sentence describing the error that occurred. This sentence is also be richly styled. In this diff the styles I enumerate are `Text` for plain text, `Emphasis` for bold text, `Code` for inline code snippets which should be rendered in a monospace font and not broken over multiple lines, `Reason` for rendering our reasons with a location, and `ReasonDesc` for only rendering the description of a reason.
3. Any number of extra locations which a user is advised to visit when debugging their error. These locations are always associated with a `Reason` from the error message. These locations, or “references,” would ideally be rendered in the IDE as hyperlinks attached to their reason in the message. In the CLI since we can’t have hyperlinks the references are rendered underneath the message.

I add a custom CLI rendering in the following diff for friendly errors, but for our JSON output of friendly errors currently I’m coercing the friendly errors into classic errors. To do this we take the primary location (1) and show it as the only location in the classic error message. Then we take the message (2) render it to a markdown string and blame the primary location with that text. Finally, we use the space for “extra” information to render all of our references (3). Each reference shows its location, displays its reference number, and shows the reason description for the reference.

This diff also introduces a few utility functions which I use in subsequent diffs.

Reviewed By: gabelevi

Differential Revision: D5989662

fbshipit-source-id: cf63906c57e8acb0e727de1506d26d7fc07d5873