Print a description of the expression AST in root use_op reasons
Summary:
In D6516207 I introduced special handling for a limited subset of reason descriptions which printed as code. This improved error messages, but the solution was very incomplete. It didn't handle all expressions and was very succeptible to slight changes in type-checking behavior.

This led to a bunch of other diffs like D6562010, D6562012, and D6562013 to help fix the inconsistencies. This approach was not scalable.

In addition, the system made it difficult and inefficient to extend the printing to other common expressions. Like literals.

This diff performs the same basic functionality as D6516207 except in a much more robust way. Instead of printing a subset of reason descriptions, we print a summary of the code directly from the AST itself. This means errors mentioning expressions are not influenced by Flow's typechecking decisions and can directly reference user code.

Reviewed By: gabelevi

Differential Revision: D6669282

fbshipit-source-id: 408b7405083e236992500a39f9b676662d8cd8ea