Better locations for references
Summary:
Both the repositioned loc in reasons and the `def_loc` are not the optimal locations for references in the friendly error message format. The best location is the _nearest annotation_. Consider:

```
const y1 = 42;
const y2: number = -42;
const y3 = y1;
const y4 = y2;
const y5: number = y1;
const y6 = y1;

(y1: empty);
(y2: empty);
(y3: empty);
(y4: empty);
(y5: empty);
(y6: empty);
```

(Added as a test in a few diffs.)

The best location to point to is `number` or `42`. Whichever comes first. To demonstrate why this location is the best, consider:

```
class A {}
class B {}

let x;

if (Math.random() > 0.5) {
  x = new A();
} else {
  x = new B();
}

(x: B);
```

Here the closest annotation for the error is `new A()`. We want to point the user there instead of saying:

```
12: (x: B);
     ^ A. This type is incompatible with
12: (x: B);
        ^ B
```

This diff gives us the information we need to do this.

Reviewed By: gabelevi

Differential Revision: D6056153

fbshipit-source-id: c9708d7e4015712e32dc57d63153c9dce57b9012