[hack] Make Lsp_helpers parameterized
Summary:
It bugs me that, while in an object oriented language it'd be easy to create a per-instance parameterized library object
```
const myHelper = new LspHelpers(A);
myHelper.foo();
```

the OCaml equivalent is more clunky:
```
let env = MyLspHelpers.make

MyLspHelper.foo env
```

It feels clunky that we have to provide both the module-qualifier `MyLspHelper` and the constructor parameter `env` at the callsite. It's nice how object-orientation lets you wrap both things up in a single identifier.

Nevertheless, that's the functional way, so that's what I'm doing in this diff.

(Previously, Lsp_helpers had stuck its head in the sand and pretended it was okay to have a global variable instead of a per-instance initialization variable. That wasn't really true, though.)

Reviewed By: arxanas

Differential Revision: D6658046

fbshipit-source-id: f7bfb2f4d84dc7d67702da6711eb425d4e68ac19