[hack] Take "request" out of Jsonrpc
Summary:
Previously, I'd written Jsonrpc.notify+response+request methods.
I had been expecting to share these three things between Flow and Hack.
Hence the desirability of moving them all into Jsonrpc.

But as it turns out, Flow couldn't use them. (That's because Flow client needs to
respond with fully-formed JSON over stdout to the editor, and
also the Flow server needs to respond with strongly-typed LSP structures
over an ocaml marshalling channel to the Flow client).

I've left in Jsonrpc.notify+response for now as helpers that Hack uses.
But this diff removes the "request" method.
The "request" method was interesting because it takes callbacks which
asynchronously handle the eventual response from the client. These
had to be parameterized on the "state" of the LSP server at the time
the response comes back. Therefore Jsonrpc itself had to be functorized.

Now that "request" has been removed, there's no longer a need for Jsonrpc
to be functorized.

I've also removed "request_showMessage" and "dismiss_showMessageRequest" from Lsp_helpers.
They also couldn't be shared with Flow because Flow's callbacks
are strongly-LSP-typed and Hack's callbacks are weakly-JSON-typed for now.
If we move Hack to be more strongly-typed, like Flow, then maybe at that time
we can put common routines back into Lsp_helpers.

Reviewed By: arxanas

Differential Revision: D6851576

fbshipit-source-id: 86060df6a8f42420ca251a3f522008ea853a58cf