Flow server --autostop
Summary:
Scenario 1: you type "flow" at the keyboard to spin up the flow server, and you expect it to keep running.

Scenario 2: you launch your code editor and open a flow project. You want the flow server to spin up automatically to power language-services, and also want it to shut down automatically once you close the project/editor.

Previously we handled Scenario2 inside Nuclide by having Nuclide track whether or not it was responsible for starting Flow; if it was, and if the user had opted into stopping Flow server once Nuclide closes the project, then when Nuclide closes the project it will issue a SIGKILL on the flow server.

This won't work with the move to LSP. That's because the editor now only interacts with the LSP server via the LSP protocol; this stuff about "responsible for starting" and "sigkill" isn't part of the protocol.

So I've adopted a new approach. Nuclide can now pass an option "--autostop" when it launches Flow (if the user has opted to stop flow upon project close). The meaning of this flag is that if the flow server was STARTED by someone who did --autostop, then when the Flow server has finished processing its last remaining ephemeral and persistent clients, then it will shut down.

Reviewed By: gabelevi

Differential Revision: D6487425

fbshipit-source-id: 8646996f81ab5e18e5c016ce5663f837bf5bcea0