Factor out "make_env" in CommandUtils
Summary:
The "env" structure is used by "connect" (which does retries) and also by "start_flow_server" (which is called by connect).

For LSP, it has a different paradigm for how to connect to and start the flow server. I'm factoring out "make_env" so that LSP can share the same start_flow_server, not via connect.

Reviewed By: gabelevi

Differential Revision: D6487478

fbshipit-source-id: 9f24593e1c633b140c8b82f66abf1a63c5de76eb