Persistent LSP connection part 1: add a flag
Summary:
This diff adds a second kind of persistent connection, an "LSP persistent connection". The diff threads this flag all the way through from when it's requested by the client, through the monitor, to the server, and in the 'client' data structure.

We'll use the flag in future diffs. For now I just want to thread it through.

The future plans... (1) if the server is still initializing, then a request for an "LSP persistent connection" will be rejected outright; requests for all other connection kinds will be accepted. (2) when the server decides to publishDiagnostics, it will publish them in LSP format for LSP persistent connections, and in the existing format for existing persistent connections.

Reviewed By: gabelevi

Differential Revision: D6508153

fbshipit-source-id: 8ce78c099de13527440363caeb80f6e9ae51f9c8