Persistent LSP connection part 2: fail if not ready
Summary:
The LSP adapter has to deal with the different phases of initialization and report them to the user:
1. if there's no flow server process at all
2. if there is a flow monitor willing to accept connections, but the flow server is still busy initializing
3. if flow server has finished initialization but is busy
4. if the flow server has finished initialization and is free

The LSP adapter might be in a "disconnected" state, where there's no connection to the flow server at all, and it polls once a second to see if the flow server has been spun up. [1]

The LSP adapter might be in a "connected" state, where there is a working connection to the flow server, and it has to display busy status [3,4]

The question is, how to implement state [2]? Should we code it as part of the "connected" state as is currently implemented for all other client types, where they complete the SocketHandshake and then just wait to become responsive? Or should we code it as part of the "disconnected" state?

I found it easier to code it as a "disconnected" state.

This diff achieves that. If the client requests a persistent LSP adapter, but the server is still initializing, then it responds to the handshake with a Server_busy error.

Reviewed By: gabelevi

Differential Revision: D6508157

fbshipit-source-id: 2c6e9d35acfc66b344866600f2d3eeac3ee66d89