Use `--file-watcher none` in `runtests.sh`
Summary:
For ages we've known that we must always call `force-recheck`
after making changes in tests, since we worry about races between the
test script and the file watcher.

This diff takes it a step further and turns off the file watcher by
default in tests. I then had to go fix the few tests which failed to
properly call force-recheck.

Reviewed By: samwgoldman

Differential Revision: D6761259

fbshipit-source-id: 49ab44329a7ad657a9446c7ebf1cb84281ab01bd