Point to full JSX element instead of just the opening element
Summary:
Before, error messages would only point to the opening tag of a JSX element when the full element was incompatible. This area can be considerably smaller then the full element size for elements with many children. However, with the introduction of alternative error message positioning options it makes more sense to higlight the full element for consistency when necessary.

This also lets me get rid of a small repositioning hack in the `PropertyCompatibility` friendly error unwrapper. (See `repos_small_reason`.)

Reviewed By: gabelevi

Differential Revision: D6729108

fbshipit-source-id: 08fa4789d0d7d81c8f909b3cd39898eb555e7c85