Fix for top-level find-refs under lazy mode
Summary:
Previously, we would only find references in files which were part of the checked set for lazy mode. For lazy filesystem mode, that is the set of files which have been touched and their transitive dependents. For lazy IDE mode, it is the set of files which have been opened along with their transitive dependents.

This adds the file with the symbol to the checked set before starting find-refs. For IDE lazy mode this is likely to be a no-op because right now, you have to start your find-refs request from the file that exports the symbol (you cannot currently find-refs from a use in a downstream file). If it's open in the IDE, it has already been added to the checked set.

This is mainly to act as a safeguard in case something's going on with IDE mode, or if people are calling find-refs from the command line even though they have an IDE open. It also makes sure find-refs works with filesystem lazy mode, though I'm not sure how much that is used. Once we support finding refs from a use, it will become more important because the file where the symbol is defined might not have been opened in the IDE.

Reviewed By: gabelevi

Differential Revision: D6877873

fbshipit-source-id: d931ab093d70eb04b8f281c6fdb95ec0b3cdb863