Add TODO comments to track friendlification progress
Summary:
80 TODOs. Let's jam.

This is more of a development tool for me. Putting it up so you can see.

Before this diff, I've converted all errors that use `use_op`s. (Excluding speculation error messages.) After this diff we have the longtail of 80 error messages that don't use `use_op`s. These 80 error messages are likely a small percent of what a user sees on a daily basis.

Reviewed By: gabelevi

Differential Revision: D6732227

fbshipit-source-id: 1f7aab1210e94a4a5fb3c9e4b56702b0c85f1af9