EValueUsedAsType, EUnexpectedTypeof, EMissingAnnotation, EUnreachable, EInvalidTypeof, and EUnsupportedImplements
Summary:
17/80

Converts `EValueUsedAsType`, `EUnexpectedTypeof`, `EMissingAnnotation`, `EUnreachable`, `EInvalidTypeof`, and `EUnsupportedImplements`.

The theme of this batch is "`typeof` and other annotation (?) errors." `EUnreachable` is definitely the odd one out. But I think it may be easier to review this one diff instead of a seperate diff just for `EUnreachable`. Maybe batches of 5-6 is still a little too hard to review.

I'm finding it hard to come up with really good error messages for these one-off errors. Perhaps its not that important? As always, better suggestions then what I iterated towards are greatly appreciated.

Reviewed By: gabelevi

Differential Revision: D6733587

fbshipit-source-id: 6cec99c4a7fd8f4e9855af016dd9b5be0fadac3b