Add reference locations as suppressable locations
Summary:
Introducing lazy mode meant that suppression comments which suppressed an error not caused by type-checking the code below the suppression (e.g. by putting a suppression over a reference) meant that unused suppression errors would pop up in undesirable places. gabelevi solved this by making unused suppression diagnostics warnings instead of errors. Another solution, which was discussed, is to ensure that all error messages point to the "correct" primary location and only allowing the suppression of that location. "Correct" meaning all references in the error are dependencies of the primary location.

I started the error message work with this in mind. Only allowing errors to be suppressed at their primary location. After a lot of work on error positioning across the team, "correctness" issues in error positioning should mostly be gone. There are likely still a few incorrectly positioned errors, but they have disappeared to the point where we could be comfortable forcing suppressions to point to the root location.

This diff allows errors in the new format to be suppressed by a comment over their reference locations. Even though there is some desire to only allow the primary location to be suppressed, currently Flow tests and repos depend on this behavior. For now, my conclusion is to allow suppressing using reference locations. In the future, we can change this behavior.

This diff is oddly positioned in my stack because the next diff I submit will change the format for lint errors. The suppressions for lint errors depend on this behavior.

Reviewed By: gabelevi

Differential Revision: D6751765

fbshipit-source-id: caa2d47cb1dc61973ff26197b72339beb71eb44d