Grab AST from the heap when possible
Summary:
In most cases we don't need to recompute the AST, since it should not be
different from what we already have on disk. So, we can just grab these ASTs
from the heap.

This change seems to lead to a 1-2 second improvement out of around 10 seconds
on the queries I've tried where there are a lot of dependents.

Reviewed By: panagosg7

Differential Revision: D6866572

fbshipit-source-id: b46a047f613b8c3a23d78d5107685cf8b36916f6