[Normalizer] Normalized type structure for user consumption
Summary:
This introduces a new simplified type structure as the target of type
normalization (`typing/type_normalizer.ml`). Most of the constructors
here are unsurprising. We focus on two points:

**Recursive types**: Consider the type:
```
{ n: number; next: { n: number; next: ... }}
```
This can be succinctly represented as the type of a variable `A` where:
```
type A = { n: number; next: A }
```
However, type `A` may not always be provided. To represent a type like this
we introduce a `Mu(v, t)` type. Here `v` is a "fresh" type variable that can
appear inside `t` and represents the type that is being defined. This type
will be serialized into a more recognizable form later on (similar to the
`type A = ...` above.

**Class types**: The second argument of `Class` will be `true` if this
is an interface type; false if it's a class type.

Reviewed By: samwgoldman

Differential Revision: D6735937

fbshipit-source-id: e4913062a6085cfb03221d1fad134686546ad4f7