Ensure directed-ness of tin_to_props/props_to_tout
Summary:
These two methods are designed to set up specific constraints, either
`tin <: props` or `props <: tout`. They do this for functional components, but
the class components' props are invariant, so we would emit `props = tin` and
`props = tout` instead.

This diff just adds a little indirection to make things go in the right
direction, which uncovered a bug (fixed earlier in the stack).

Reviewed By: avikchaudhuri

Differential Revision: D6920647

fbshipit-source-id: 05eef5a241a45120716f88dd2e7624d7a4baa861