Special-case spreads of maybe/optional
Summary:
In D6903635, I fixed a stall in the spread implementation. This fix caused a
large number of new errors. Some new errors are expected. The great number was
partially because the buggy implementation for spreading maybe/optional was
convenient.

In particular, when we spread `?any` or `?Object`, the result should be `any`
or `Object` respectively. This worked before because the union-like type would
be split up, the null/void cases would do nothing, and the any-like case would
propagate the expected type.

After D6903635, the null/void case and the any-like case propagate separately.
This change ensures that we don't split up the union.

Reviewed By: avikchaudhuri

Differential Revision: D6906174

fbshipit-source-id: 53ad13344266992d8edff3ad0293c60a17f12658