[Normalizer] State and result monad composition
Summary:
This provides a monad that is the composition of a state and a
result monad. Here we forego defining a general state monad
transformer that would wrap another monad. Instead, we directly
define a monad that is essentially the result of the composition.

This monad is useful for stateful computations that may return
an erroneous state (e.g. normalizer).

Reviewed By: samwgoldman

Differential Revision: D6772025

fbshipit-source-id: 6155a948a5b3761b469197f12843fcdbbe69f479