[hack] Add type pretty-printing
Summary:
Add a pretty-printer for the `Typing_defs.ty` data structure, and use it in the pretty-printer for `Tast.saved_env`.

I would automatically generate this using the ppx_deriving show plugin, but it does not support GADTs. It does, however, produce output for GADTs which can be made to compile by manually adding locally abstract types, so that's what I've done here. I've also done some cleanup of the generated code to make it easier to maintain as we make changes to Typing_defs in the future.

Reviewed By: dlreeves

Differential Revision: D6968032

fbshipit-source-id: b86fcaa2d70770a1cc1844bfc857cd7d06660d78