aggregate use->def map api
Summary:
Consider the problem of transforming a map fd: def -> R to a map fu: use -> R, such that whenever
fd(d) = r, we have fu(u) = r for all uses u of def d.

Computing uses of individual defs is costly, since the entire scope tree needs to be
traversed. Instead, we can traverse the entire scope tree once to compute a map ud: use -> def, and
use it to define fu as the composition of fd and ud.

Reviewed By: samwgoldman

Differential Revision: D6986808

fbshipit-source-id: 1f3a26733bd8ee913ec4ee326fa8258fcaf4d1cc