Object type find-refs
Summary:
This diff extends the logic used for class member find-refs to also support
ordinary objects.

```
type Foo = {bar: string};
const x: Foo = {bar: ''};
x.bar;
```

In this example, find-refs on the `bar` in line 1 or 3 will return the locations
for both of those `bar`'s. Currently, the `bar` in the object literal will not
be included but I have explored how to include it and I have an approach which
is feasible. I'll implement that in the coming days.

This diff also causes some locations to be provided more than once (see the test
output). I know why this is happening and it's harmless but I'll still fix it in
a follow-up diff.

Reviewed By: panagosg7

Differential Revision: D6983491

fbshipit-source-id: 7bfd3bf7bf225e9eba9ccac4293e19ee79dd0a30