robust enum representation opt
Summary:
Unions of literal types enjoy a specialized representation as a "EnumSet": a set of literal
types. This representation enables quick membership checks (resulting in success or failure) without
having to go through the expensive process of speculative matching.

Unfortunately, the representation of a union was picked at the time of creation, which meant that if
any part of the union was unresolved, we would not pick a specialized representation. In practice,
the addition of a single indirection in a large enum would therefore result in catastrophic
deoptimization.

This diff makes it possible to improve the representation after creation, by using an OCaml ref to
model the representation state. Before we begin speculative matching, we we try to optimize the
representation (unless it's already optimized) and take advantage of the optimization.

The representation state machine is still preliminary, and will be addressed in immediately
following work.

Reviewed By: samwgoldman

Differential Revision: D6940957

fbshipit-source-id: 2b0dc150bd2c2a6d36c9a73d55a84b5719b5fddc