Add error message text rendering
Summary:
Adds rendering for error message text right below the header. leebyron made the excellent observation that it makes more sense for the error message text to come //before// the code frame.

The rendered text when color is disabled is not different from our fallback rendering of the error message text. Except that we line break this error message text. We line break at the same message width that we used in the previous diff for error message headers. The width defaults to `max(120, terminalWidth)` but can be configured with the new `--message-width` flag.

I tried to extensively document the line break logic. Roughly it occurs in three phases. In the first phase we turn our `Friendly.message` data structure into an array of styled strings which are annotated with either `true` or `false` which represents whether or not they are breakable. Styled strings with `true` may be broken at space characters (` `). Styled strings with `false` may not be broken at all.

In our program, `Text` is allowed to be broken, but inline `Code` is not allowed to be broken. In a future diff when I add references (e.g. "number [1]") we will not allow the space between the reference descriptor ("number") and the reference identifier ("[1]") to be broken across multiple lines.

In the second phase we break our styled strings into words. Breakable strings are broken out into words whever we see a space character (" "). Unbreakable strings are kept as one word.

Finally, in the third phase we take our words and arrange them into lines. If a word cannot fit onto the current line then we put that word on the next line.

In this logic, I don't handle orphans gracefully. e.g.

```
Hello world. My name is
Doug.
```

In this message, the word "Doug" is an orphan. A more aesthetically pleasing rendering would be:

```
Hello world. My name
is Doug.
```

Unfortunately, this is not the most complicated logic for the new CLI renderer.

Likewise, all the errors in our test suite changed which may make this diff difficult to review. I added new tests in `cli_renderer_*` and attached photos in the Test Plan which should help you review.

Reviewed By: gabelevi

Differential Revision: D6836788

fbshipit-source-id: ddcc697aeb79bde2f5ea504455bed96358ed7e55