Group error messages with the same root operation into a single chunk
Summary:
This diff implements what is, in my mind, one of the most significant improvements of the error message work. Because Flow's type system is structural, it's possible that one root operation could generate multiple incompatibilities. For instance:

```
f({a: 1, b: 2});

function f(args: {a: string, b: string}) {}
```

Here, both `1` is incompatible with `string` and `2` is incompatible with `string`. But both of these errors are generated by the same root operation.

This example is fairly trivial, but it could be common for a user to take one third-party type and pass it to a function that expects another third-party type. These types could have multiple incompatibilities. However, for the user, there is only one root cause.

By grouping errors together by root cause, we save space and improve the user's recognition of what went wrong in their program.

fishythefish has observed a couple places in previous diffs where this behavior would be useful.

I realize that this diff may be a little difficult to review considering I removed the legacy error message rendering in the previous diff. Not being able to see locations makes it harder to determine whether grouping is happening correctly or not. Let me know if you'd like me to improve the reviewability of this diff by somehow showing the error locations.

In the error diff you will see a couple of instances where the same message is listed twice. e.g.

```
Cannot cast `x` to `T` because:
- number is incompatible with string.
- number is incompatible with string.
```

This is because the code which generates the error has two different `number`s at different locations. So the errors were not de-duped. When I introduce references, this will be eaiser to visualize. Since the text will instead look like:

```
Cannot cast `x` to `T` because:
- number [1] is incompatible with string [2].
- number [3] is incompatible with string [2].
```

When Unicode is enabled we use the bullet character [U+2022](http://graphemica.com/%E2%80%A2).

The way grouping is implemented is that we depend on `ErrorSet` sorting to put error messages with the same root operation next to each other. Then when we print the messages, we fold over the `ErrorSet` grouping together //adjacent// error messages with the same root operation. Thanks to the sort, these messages will always be adjacent.

Reviewed By: gabelevi

Differential Revision: D6836971

fbshipit-source-id: 971670127d66bba256bbbcdce953fe3b60459fcb