Implement colored code frames
Summary:
This diff is pretty complex. We use the `layout_references` function added in the previous diff to render all the code frames necessary to understand an error with highlighted references. See the test plan for images.

Most of the logic for rendering code frames is in `print_code_frames_friendly`. We do some preparation work before printing the code frames in `layout_friendly_error_group`. The steps for rendering code frames are roughly as follows:

1. `layout_friendly_error_group`:
  1. Extract references from the error message text.
  2. Add "shadow references" (references with a negative identifier) with custom colors for the root location (white) and primary locations (red).
  3. Layout the references and shadow references with `layout_references`.
2. `print_code_frames_friendly`:
  1. Get a list of all the locs we want to display by file. The root loc is expanded by 3 lines in both directions.
  2. Merge intersecting locs and locs within 3 lines of one another. (This is why `Loc.lines_intersect` was added.)
  3. Compute the reference numbers we will put on each line. Through this process, determine what the gutter width needs to be. (`file_line_references`)
  4. Read the lines of code for each loc we computed earlier.
  5. Loop through the lines of code for each loc and `tags` from `layout_references` together. Whenever we see a tag inside our line of code then split the line at the tag's position and set the appropriate styling as determined by previous tags.
  6. Add the line number and references from `file_line_references` to the start of the line which was styled in the previous step.
  7. At this point, combining each individual line we read from source code files creates a "code frame." Concatenate all of our code frames together and add headers for code frames in files other then the root file.

This diff also adds a new parameter to `layout_references`: `custom_colors`. This parameter allows us to give certain locations a special meaning. References with a custom color do not contribute to the color ranking algorithm other references take part in.

While trying to decide how to render this design without color, I couldn't think of a good way to pack the high information density we get with color into an error message without color. So I'm going to create a completely different function to render code frames without color. This function will spread out information instead of packing it together.

Unfortunately, this means we can't test this design in our test suite. Since our test suite is not color enabled. Because of this, I tried to be extra rigorous in writing tests for `cli_renderer_color`.

samwgoldman made the excellent recommendation to test this design against colorblind engineers. When I distribute a beta to a few people, I will try to get feedback from someone who is colorblind. Although, even if this design does not work too well for the colorblind, these individuals have a lot of options to improve contrast. Terminal colors can be configured and Flow will provide a `--color=never` version to remove the need for color entirely.

Reviewed By: gabelevi

Differential Revision: D6837107

fbshipit-source-id: a61487c7d2fe19d283456c274a3d5939031e4e13