Add function to layout reference locations
Summary:
In this diff, I add the `layout_references` function. Its effects are almost un-noticeable in the error diff. The main change to the error diff is the addition of reference numbers. e.g. `[1]` and `[2]`. However, the role of `layout_references` is to compute which substrings in our source text need to be higlighted and with what color. We must compute the color of each reference before rendering the source text so that we may use the same color for reference numbers.

I attempt to explain the what and why of `layout_references` in a comment within the source text, so I will leave out the explanation in this diff description.

At a high-level, `layout_references` takes a map of references and returns two data structures:

1. A map recording the color for each reference. This map should have the same keys as the references map `layout_references` was passed.
2. A list of "tags" which can be used when rendering our source frame to procedurally higlight the correct substrings in our source text.

`layout_references` is likely more inefficient then it could be. I prioritized correctness and implementation grokability over efficiency where the tradeoff appeared.

For instance, we need to insert every reference into our "tags" list. The tags list is ordered by source positions, so we could represent the tags list as some binary tree for efficient insertions. Instead, `layout_references` does a linear scan of the tags linked list to determine where to insert each new reference location.

Performing a linear scan allows us to keep track of all the references in the tags list that have previously been added. Allowing us to establish the correct color for the reference we are inserting. Using a binary-tree, or some other data structure, would require us to record this information in the data structure since we would not scan over all the information.

I'm not too concerned about inefficiencies in `layout_references`. `layout_references` is not in a hot path and I never expect there to be more then 100 references in a single file.

Reviewed By: gabelevi

Differential Revision: D6837040

fbshipit-source-id: 981fb224da944fe82ffd8a19644afdac40508411